Split approvals unit tests in 4 files
